refactor message generation (ros2/ros2#48)

+++ rosidl_typesupport_introspection_cpp/__init__.py
@@ -16,6 +16,7 @@ import em
 from io import StringIO
 import os
 
+from rosidl_cmake import convert_camel_case_to_lower_case_underscore
 from rosidl_cmake import extract_message_types
 from rosidl_parser import parse_message_file
 from rosidl_parser import parse_service_file
@@ -24,38 +25,42 @@ from rosidl_parser import validate_field_types
 
 def generate_cpp(pkg_name, ros_interface_files, deps, output_dir, template_dir):
     mapping_msgs = {
-        os.path.join(template_dir, 'msg_TypeSupport_Introspection.cpp.template'):
-        '%s_TypeSupport_Introspection.cpp',
+        os.path.join(template_dir, 'msg__type_support.cpp.template'):
+        '%s__type_support.cpp',
     }
 
     mapping_srvs = {
-        os.path.join(template_dir, 'srv_ServiceTypeSupport_Introspection.cpp.template'):
-        '%s_ServiceTypeSupport_Introspection.cpp',
+        os.path.join(template_dir, 'srv__type_support.cpp.template'):
+        '%s__type_support.cpp',
     }
 
     for template_file in mapping_msgs.keys():
-        assert(os.path.exists(template_file))
+        assert os.path.exists(template_file), 'Could not find template: ' + template_file
 
     for template_file in mapping_srvs.keys():
-        assert(os.path.exists(template_file))
-
-    try:
-        os.makedirs(output_dir)
-    except FileExistsError:
-        pass
+        assert os.path.exists(template_file), 'Could not find template: ' + template_file
 
     known_msg_types = extract_message_types(pkg_name, ros_interface_files, deps)
 
+    functions = {
+        'get_header_filename_from_msg_name': convert_camel_case_to_lower_case_underscore,
+    }
+
     for ros_interface_file in ros_interface_files:
-        filename, extension = os.path.splitext(ros_interface_file)
+        extension = os.path.splitext(ros_interface_file)[1]
+        subfolder = os.path.basename(os.path.dirname(ros_interface_file))
         if extension == '.msg':
             spec = parse_message_file(pkg_name, ros_interface_file)
             validate_field_types(spec, known_msg_types)
             for template_file, generated_filename in mapping_msgs.items():
-                generated_file = os.path.join(output_dir, generated_filename % spec.base_type.type)
+                generated_file = os.path.join(
+                    output_dir, subfolder, generated_filename %
+                    convert_camel_case_to_lower_case_underscore(spec.base_type.type))
 
                 try:
                     output = StringIO()
+                    data = {'spec': spec, 'subfolder': subfolder}
+                    data.update(functions)
                     # TODO reuse interpreter
                     interpreter = em.Interpreter(
                         output=output,
@@ -63,13 +68,14 @@ def generate_cpp(pkg_name, ros_interface_files, deps, output_dir, template_dir):
                             em.RAW_OPT: True,
                             em.BUFFERED_OPT: True,
                         },
-                        globals={'spec': spec},
+                        globals=data,
                     )
                     interpreter.file(open(template_file))
                     content = output.getvalue()
                     interpreter.shutdown()
                 except Exception:
-                    os.remove(generated_file)
+                    if os.path.exists(generated_file):
+                        os.remove(generated_file)
                     raise
 
                 # only overwrite file if necessary
@@ -77,6 +83,10 @@ def generate_cpp(pkg_name, ros_interface_files, deps, output_dir, template_dir):
                     with open(generated_file, 'r') as h:
                         if h.read() == content:
                             continue
+                try:
+                    os.makedirs(os.path.dirname(generated_file))
+                except FileExistsError:
+                    pass
                 with open(generated_file, 'w') as h:
                     h.write(content)
 
@@ -84,10 +94,14 @@ def generate_cpp(pkg_name, ros_interface_files, deps, output_dir, template_dir):
             spec = parse_service_file(pkg_name, ros_interface_file)
             validate_field_types(spec, known_msg_types)
             for template_file, generated_filename in mapping_srvs.items():
-                generated_file = os.path.join(output_dir, generated_filename % spec.srv_name)
+                generated_file = os.path.join(
+                    output_dir, subfolder, generated_filename %
+                    convert_camel_case_to_lower_case_underscore(spec.srv_name))
 
                 try:
                     output = StringIO()
+                    data = {'spec': spec}
+                    data.update(functions)
                     # TODO reuse interpreter
                     interpreter = em.Interpreter(
                         output=output,
@@ -95,13 +109,14 @@ def generate_cpp(pkg_name, ros_interface_files, deps, output_dir, template_dir):
                             em.RAW_OPT: True,
                             em.BUFFERED_OPT: True,
                         },
-                        globals={'spec': spec},
+                        globals=data,
                     )
                     interpreter.file(open(template_file))
                     content = output.getvalue()
                     interpreter.shutdown()
                 except Exception:
-                    os.remove(generated_file)
+                    if os.path.exists(generated_file):
+                        os.remove(generated_file)
                     raise
 
                 # only overwrite file if necessary
@@ -109,6 +124,10 @@ def generate_cpp(pkg_name, ros_interface_files, deps, output_dir, template_dir):
                     with open(generated_file, 'r') as h:
                         if h.read() == content:
                             continue
+                try:
+                    os.makedirs(os.path.dirname(generated_file))
+                except FileExistsError:
+                    pass
                 with open(generated_file, 'w') as h:
                     h.write(content)
